refactor: move PlotDataResult and plot TypedDicts to shared/boundaries/
PlotDataResult and 6 related TypedDicts (SpectrogramResult,
MatchedAmpVsSpeedSeries, FreqVsSpeedByFindingSeries, AmpVsPhaseRow,
PhaseSegmentOut, PhaseBoundary) were defined in use_cases/diagnostics/plots.py
but referenced by shared/boundaries/analysis_payload.py via a reverse
TYPE_CHECKING import.

These are boundary payload shapes, not business logic. Moving them to
analysis_payload.py eliminates the shared→use_cases dependency.

Closes #745
